question_3/requirements.txt
tabulate==0.9.0
PyYAML==6.0.1